Codewarrior is very picky with the floating point constants passed to templated routines.  Making sure all parameters are floats.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7051 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/emit/object/emit.cpp
@@ -140,7 +140,7 @@ void csEmitSphere::GetValue(csVector3& value, csVector3 & /*given*/)
   // is the dist*dist(*pi..) in size. Taking random min..max will
   // cause an uneven spread of points in the sphere.
   float sqdist = GetRandomFloat( min*min*min, max*max*max );
-  float dist = pow(sqdist, 1./3.);
+  float dist = pow(sqdist, (float)(1./3.));
   value.Set(dist, 0, 0);
   float rotz_open = 2.0 * PI * (rand() / (1.0+RAND_MAX));
   csZRotMatrix3 openrot(rotz_open);
